+++ package/recalbox-system/rpi4/pre-upgrade.sh
@@ -39,3 +39,12 @@ fi
 sed -i \
     -e '/^\(audio.device=\).*/{s//\1alsa_card.1:hdmi-output-0/;:a;n;ba;q}' \
     -e '$aaudio.device=alsa_card.1:hdmi-output-0' /recalbox/share/system/recalbox.conf
+
+# Force VSync on PPSSPP
+# If file is not here, we let the user take care of it 
+PPSSPPCONFIGILE="/recalbox/share/system/configs/ppsspp/PSP/SYSTEM/ppsspp.ini"
+if [ -f "${PPSSPPCONFIGILE}" ]; then
+  sed -i \
+    -e 's/^VSyncInterval\s*=.*/VSyncInterval = True/' \
+    "${PPSSPPCONFIGILE}"
+fi